changed return type of getNewFrame to float[] for efficiency, including changing internal data type to float[]. This change could affect users!  It reduces memory usage by a factor of two and should speed up computes by ~3x on the values.
git-svn-id: https://svn.code.sf.net/p/jaer/code/jAER/trunk@6014 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/jaer/projects/davis/frames/DAVISAVIWriter.java
@@ -93,7 +93,7 @@ public class DAVISAVIWriter extends EventFilter2D implements PropertyChangeListe
     @Override
     public void propertyChange(PropertyChangeEvent evt) {
         if (aviOutputStream != null && evt.getPropertyName() == ApsFrameExtractor.EVENT_NEW_FRAME) {
-            double[] frame = apsFrameExtractor.getNewFrame();
+            float[] frame = apsFrameExtractor.getNewFrame();
 
             BufferedImage bufferedImage = new BufferedImage(chip.getSizeX(), chip.getSizeY(), BufferedImage.TYPE_3BYTE_BGR);
             WritableRaster raster = bufferedImage.getRaster();
